split semver
