+++ lib/manager/npm/lookup.js
@@ -1,5 +1,4 @@
-// const datasource = require('../../datasource/npm');
-const versioning = require('../../versioning/semver.js');
+const versioning = require('../../versioning/semver');
 const moment = require('moment');
 
 const {
@@ -102,7 +101,6 @@ function lookupUpdates(dependency, config) {
       fromVersion = maxSatisfyingVersion(allVersions, currentVersion);
     }
   } else {
-    // console.log('currentVersion is pinned');
     fromVersion = currentVersion;
   }
   // Leave only versions greater than current
