+++ lib/versioning/semver/range.js
@@ -1,47 +1,16 @@
 const semver = require('semver');
-const stable = require('semver-stable');
 const semverUtils = require('semver-utils');
 
-const { is: isStable } = stable;
+const { major, minor, valid } = semver;
+const { parseRange } = semverUtils;
 
-const { parseRange, parse: parseVersion } = semverUtils;
-
-const {
-  compare: semverSort,
-  gt,
-  gtr,
-  lt,
-  ltr,
-  intersects: intersectsSemver,
-  maxSatisfying: maxSatisfyingVersion,
-  minSatisfying: minSatisfyingVersion,
-  minor: getMinor,
-  patch: getPatch,
-  satisfies: matchesSemver,
-  valid: isPinnedVersion,
-} = semver;
-
-const padRange = range => range + '.0'.repeat(3 - range.split('.').length);
-
-const getMajor = input => {
-  const version = isPinnedVersion(input) ? input : padRange(input);
-  return semver.major(version);
+module.exports = {
+  rangify,
 };
-
-const isRange = input => isValidSemver(input) && !isPinnedVersion(input);
-
-// If this is left as an alias, inputs like "17.04.0" throw errors
-const isValidSemver = input => semver.validRange(input);
-
-const isLessThan = (version, base) =>
-  isPinnedVersion(base) ? lt(version, base) : ltr(version, base);
-
-const isGreaterThan = (version, base) =>
-  isPinnedVersion(base) ? gt(version, base) : gtr(version, base);
 
 function rangify(config, currentVersion, fromVersion, toVersion) {
   const { rangeStrategy } = config;
-  if (rangeStrategy === 'pin' || isPinnedVersion(currentVersion)) {
+  if (rangeStrategy === 'pin' || valid(currentVersion)) {
     return toVersion;
   }
   const parsedRange = parseRange(currentVersion);
@@ -90,28 +59,28 @@ function rangify(config, currentVersion, fromVersion, toVersion) {
     return null;
   }
   if (element.operator === '^') {
-    if (fromVersion && getMajor(toVersion) === getMajor(fromVersion)) {
-      if (getMajor(toVersion) === 0) {
-        if (getMinor(toVersion) === 0) {
+    if (fromVersion && major(toVersion) === major(fromVersion)) {
+      if (major(toVersion) === 0) {
+        if (minor(toVersion) === 0) {
           return `^${toVersion}`;
         }
-        return `^${getMajor(toVersion)}.${getMinor(toVersion)}.0`;
+        return `^${major(toVersion)}.${minor(toVersion)}.0`;
       }
       return `^${toVersion}`;
     }
-    return `^${getMajor(toVersion)}.0.0`;
+    return `^${major(toVersion)}.0.0`;
   }
   if (element.operator === '~') {
-    return `~${getMajor(toVersion)}.${getMinor(toVersion)}.0`;
+    return `~${major(toVersion)}.${minor(toVersion)}.0`;
   }
   if (element.operator === '<=') {
     let res;
     if (element.patch) {
       res = `<=${toVersion}`;
     } else if (element.minor) {
-      res = `<=${getMajor(toVersion)}.${getMinor(toVersion)}`;
+      res = `<=${major(toVersion)}.${minor(toVersion)}`;
     } else {
-      res = `<=${getMajor(toVersion)}`;
+      res = `<=${major(toVersion)}`;
     }
     if (currentVersion.includes('<= ')) {
       res = res.replace('<=', '<= ');
@@ -121,14 +90,14 @@ function rangify(config, currentVersion, fromVersion, toVersion) {
   if (element.operator === '<') {
     let res;
     if (currentVersion.endsWith('.0.0')) {
-      const newMajor = getMajor(toVersion) + 1;
+      const newMajor = major(toVersion) + 1;
       res = `<${newMajor}.0.0`;
     } else if (element.patch) {
       res = `<${semver.inc(toVersion, 'patch')}`;
     } else if (element.minor) {
-      res = `<${getMajor(toVersion)}.${getMinor(toVersion) + 1}`;
+      res = `<${major(toVersion)}.${minor(toVersion) + 1}`;
     } else {
-      res = `<${getMajor(toVersion) + 1}`;
+      res = `<${major(toVersion) + 1}`;
     }
     if (currentVersion.includes('< ')) {
       res = res.replace('<', '< ');
@@ -138,34 +107,14 @@ function rangify(config, currentVersion, fromVersion, toVersion) {
   if (!element.operator) {
     if (element.minor) {
       if (element.minor === 'x') {
-        return `${getMajor(toVersion)}.x`;
+        return `${major(toVersion)}.x`;
       }
       if (element.patch === 'x') {
-        return `${getMajor(toVersion)}.${getMinor(toVersion)}.x`;
+        return `${major(toVersion)}.${minor(toVersion)}.x`;
       }
-      return `${getMajor(toVersion)}.${getMinor(toVersion)}`;
+      return `${major(toVersion)}.${minor(toVersion)}`;
     }
-    return `${getMajor(toVersion)}`;
+    return `${major(toVersion)}`;
   }
   return toVersion;
 }
-
-module.exports = {
-  getMajor,
-  getMinor,
-  getPatch,
-  intersectsSemver,
-  isGreaterThan,
-  isLessThan,
-  isRange,
-  isStable,
-  isValidSemver,
-  isPinnedVersion,
-  matchesSemver,
-  maxSatisfyingVersion,
-  minSatisfyingVersion,
-  parseRange,
-  parseVersion,
-  rangify,
-  semverSort,
-};
